# Span aware repetition parsers