--- a/src/Parsley.Tests/IntegrationTests/Json/Json.cs
+++ b/src/Parsley.Tests/IntegrationTests/Json/Json.cs
@@ -92,7 +92,7 @@ public class Json
     {
         get
         {
-            var Digits = OneOrMore(IsDigit, "0..9");
+            var Digits = OneOrMore(IsDigit, "0..9", span => span.ToString());
 
             return
                 from index in Index<char>()
@@ -145,7 +145,7 @@ public class Json
 
             var unicodeEscapeCharacters =
                 from u in Single('u', "unicode escape sequence")
-                from unicodeDigits in Repeat(IsLetterOrDigit, 4, "4 unicode digits")
+                from unicodeDigits in Repeat(IsLetterOrDigit, 4, "4 unicode digits", span => span.ToString())
                 select char.ConvertFromUtf32(
                     int.Parse(
                         unicodeDigits,
@@ -158,7 +158,7 @@ public class Json
                 select unescaped;
 
             var literalCharacters =
-                OneOrMore(c => c != '"' && c != '\\', "non-quote, not-slash character");
+                OneOrMore((char c) => c != '"' && c != '\\', "non-quote, not-slash character", span => span.ToString());
 
             return
                 from index in Index<char>()