--- a/src/Parsley/Grammar.Repetition.cs
+++ b/src/Parsley/Grammar.Repetition.cs
@@ -176,6 +176,11 @@ partial class Grammar
 
     public static Parser<char, string> ZeroOrMore(Func<char, bool> test)
     {
+        return ZeroOrMore(test, span => span.ToString());
+    }
+
+    static Parser<char, TValue> ZeroOrMore<TValue>(Func<char, bool> test, SpanFunc<char, TValue> selector)
+    {
         return (ReadOnlySpan<char> input, ref int index, out bool succeeded, out string? expectation) =>
         {
             var length = input.CountWhile(index, test);
@@ -187,39 +192,49 @@ partial class Grammar
 
                 expectation = null;
                 succeeded = true;
-                return slice.ToString();
+                return selector(slice);
             }
 
             expectation = null;
             succeeded = true;
-            return "";
+            return selector(ReadOnlySpan<char>.Empty);
         };
     }
 
     public static Parser<char, string> OneOrMore(Func<char, bool> test, string name)
     {
+        return OneOrMore(test, name, span => span.ToString());
+    }
+
+    static Parser<char, TValue> OneOrMore<TValue>(Func<char, bool> test, string name, SpanFunc<char, TValue> selector)
+    {
         return (ReadOnlySpan<char> input, ref int index, out bool succeeded, out string? expectation) =>
         {
             var length = input.CountWhile(index, test);
 
             if (length > 0)
             {
-                var span = input.Slice(index, length);
+                var slice = input.Slice(index, length);
                 index += length;
 
                 expectation = null;
                 succeeded = true;
-                return span.ToString();
+                return selector(slice);
             }
 
             expectation = name;
             succeeded = false;
-            return null;
+            return default;
         };
     }
 
     public static Parser<char, string> Repeat(Func<char, bool> test, int count, string name)
     {
+        return Repeat(test, count, name, span => span.ToString());
+    }
+
+    static Parser<char, TValue> Repeat<TValue>(Func<char, bool> test, int count, string name, SpanFunc<char, TValue> selector)
+    {
         if (count <= 1)
             throw new ArgumentException(
                 $"{nameof(Repeat)} requires the given count to be > 1.", nameof(count));
@@ -235,16 +250,21 @@ partial class Grammar
 
                 expectation = null;
                 succeeded = true;
-                return slice.ToString();
+                return selector(slice);
             }
 
             expectation = name;
             succeeded = false;
-            return null;
+            return default;
         };
     }
 
     public static Parser<TItem, IReadOnlyList<TItem>> Repeat<TItem>(Func<TItem, bool> test, int count, string name)
+    {
+        return Repeat<TItem, IReadOnlyList<TItem>>(test, count, name, span => span.ToArray());
+    }
+
+    static Parser<TItem, TValue> Repeat<TItem, TValue>(Func<TItem, bool> test, int count, string name, SpanFunc<TItem, TValue> selector)
     {
         if (count <= 1)
             throw new ArgumentException(
@@ -261,16 +281,21 @@ partial class Grammar
 
                 expectation = null;
                 succeeded = true;
-                return slice.ToArray();
+                return selector(slice);
             }
 
             expectation = name;
             succeeded = false;
-            return null;
+            return default;
         };
     }
 
     public static Parser<TItem, IReadOnlyList<TItem>> ZeroOrMore<TItem>(Func<TItem, bool> test)
+    {
+        return ZeroOrMore<TItem, IReadOnlyList<TItem>>(test, span => span.ToArray());
+    }
+
+    static Parser<TItem, TValue> ZeroOrMore<TItem, TValue>(Func<TItem, bool> test, SpanFunc<TItem, TValue> selector)
     {
         return (ReadOnlySpan<TItem> input, ref int index, out bool succeeded, out string? expectation) =>
         {
@@ -283,16 +308,21 @@ partial class Grammar
 
                 expectation = null;
                 succeeded = true;
-                return slice.ToArray();
+                return selector(slice);
             }
 
             expectation = null;
             succeeded = true;
-            return Array.Empty<TItem>();
+            return selector(ReadOnlySpan<TItem>.Empty);
         };
     }
 
     public static Parser<TItem, IReadOnlyList<TItem>> OneOrMore<TItem>(Func<TItem, bool> test, string name)
+    {
+        return OneOrMore<TItem, IReadOnlyList<TItem>>(test, name, span => span.ToArray());
+    }
+
+    static Parser<TItem, TValue> OneOrMore<TItem, TValue>(Func<TItem, bool> test, string name, SpanFunc<TItem, TValue> selector)
     {
         return (ReadOnlySpan<TItem> input, ref int index, out bool succeeded, out string? expectation) =>
         {
@@ -300,17 +330,17 @@ partial class Grammar
 
             if (length > 0)
             {
-                var span = input.Slice(index, length);
+                var slice = input.Slice(index, length);
                 index += length;
 
                 expectation = null;
                 succeeded = true;
-                return span.ToArray();
+                return selector(slice);
             }
 
             expectation = name;
             succeeded = false;
-            return null;
+            return default;
         };
     }
 }