--- a/src/Parsley/SpanFunc.cs
+++ b/src/Parsley/SpanFunc.cs
@@ -1,0 +1,3 @@
+namespace Parsley;
+
+public delegate TResult SpanFunc<TItem, out TResult>(ReadOnlySpan<TItem> input);