This improves performance around low-level repetition parsers which recognize a slice of items from the original input span, such as recognizing a contiguous series of digit characters representing an integer.

Before this change, the discovered slice would be realized into an allocated string (if the original input items are `char`), or an allocated array (if the original input items are not `char`). Even if the caller's intention was to then pass these strings/arrays to a span-accepting method for further processing, they would incur the cost of the intermediate allocation.

With this change, the user instead provides a lambda expression which operates directly on the discovered span of matched items, giving them control over whether to allocate via ToString, ToArray, or perform some other transformation such as calling the span-accepting overloads of int.Parse without ever having to 'realize' an intermediate string.

At first this did not seem possible, as the tempting thing would be to offer a parameter of type `Func<ReadOnlySpan<TItem>>, TResult>`, which is a compiler error. `ReadOnlySpan<T>` is a `ref struct`, which can never appear as a generic type parameter. The subtlety is that `ref struct` types cannot *appear as a generic type parameter*, but they **can** *appear as inputs and outputs of custom delegate types* so long as they don't need to appear in the angle brackets in the process. Where `Func` would break the rule for span inputs:

```cs
public delegate TResult Func<in T, out TResult>(T input);
```

Our custom `SpanFunc` satisfies all the rules:

```cs
public delegate TResult SpanFunc<TItem, out TResult>(ReadOnlySpan<TItem> input);
```

# Before

|           Method |     sample |        Mean | Ratio |    Gen 0 |   Gen 1 |   Gen 2 | Allocated |
|----------------- |----------- |------------:|------:|---------:|--------:|--------:|----------:|
| System.Text.Json |  Recursion |   722.74 us |  1.00 |  42.9688 | 41.0156 | 41.0156 |    145 KB |
|  Newtonsoft.Json |  Recursion |   413.05 us |  0.57 |  37.5977 | 12.6953 |       - |    181 KB |
|          **Parsley** |  Recursion |   **630.76 us** |  0.87 |  31.2500 |  0.9766 |       - |    **131 KB** |
|           Pidgin |  Recursion |   551.00 us |  0.76 |  55.6641 |  1.9531 |       - |    229 KB |
|                  |            |             |       |          |         |         |           |
| System.Text.Json | Repetition |   338.95 us |  1.00 |  24.9023 |  5.8594 |       - |    103 KB |
|  Newtonsoft.Json | Repetition | 1,140.12 us |  3.36 | 103.5156 | 50.7813 |       - |    641 KB |
|          **Parsley** | Repetition |   **686.46 us** |  2.03 |  62.5000 | 27.3438 |       - |    **337 KB** |
|           Pidgin | Repetition | 1,897.00 us |  5.60 |  56.6406 | 25.3906 |       - |    311 KB |
|                  |            |             |       |          |         |         |           |
| System.Text.Json |    Typical |    19.02 us |  1.00 |   1.3733 |       - |       - |      6 KB |
|  Newtonsoft.Json |    Typical |    54.64 us |  2.87 |   9.7656 |  1.1597 |       - |     40 KB |
|          **Parsley** |    Typical |    **38.74 us** |  2.04 |   5.7983 |       - |       - |     **24 KB** |
|           Pidgin |    Typical |    96.94 us |  5.10 |   5.0049 |       - |       - |     21 KB |

# After

|           Method |     sample |        Mean | Ratio |    Gen 0 |   Gen 1 |   Gen 2 | Allocated |
|----------------- |----------- |------------:|------:|---------:|--------:|--------:|----------:|
| System.Text.Json |  Recursion |   729.47 us |  1.00 |  42.9688 | 41.0156 | 41.0156 |    145 KB |
|  Newtonsoft.Json |  Recursion |   417.03 us |  0.57 |  37.5977 | 12.6953 |       - |    181 KB |
|          **Parsley** |  Recursion |   **597.92 us** |  0.82 |  31.2500 |  0.9766 |       - |    **131 KB** |
|           Pidgin |  Recursion |   559.35 us |  0.77 |  55.6641 |  1.9531 |       - |    229 KB |
|                  |            |             |       |          |         |         |           |
| System.Text.Json | Repetition |   339.81 us |  1.00 |  24.9023 |  5.8594 |       - |    103 KB |
|  Newtonsoft.Json | Repetition | 1,181.44 us |  3.48 | 103.5156 | 50.7813 |       - |    641 KB |
|          **Parsley** | Repetition |   **680.93 us** |  2.00 |  58.5938 | 28.3203 |       - |    **326 KB** |
|           Pidgin | Repetition | 1,905.68 us |  5.61 |  56.6406 | 25.3906 |       - |    311 KB |
|                  |            |             |       |          |         |         |           |
| System.Text.Json |    Typical |    18.99 us |  1.00 |   1.3733 |       - |       - |      6 KB |
|  Newtonsoft.Json |    Typical |    54.45 us |  2.87 |   9.7656 |  0.8545 |       - |     40 KB |
|          **Parsley** |    Typical |    **38.95 us** |  2.05 |   5.6763 |  0.1221 |       - |     **23 KB** |
|           Pidgin |    Typical |   101.62 us |  5.54 |   5.0049 |       - |       - |     20 KB |

The measureable improvement here is in the Repetition example, where we encountere and parse int values nested within 256 consecutive JSON objects.